Refactor XmlApi class ITR-95
Extract the post related methods to a CheckPostXml middleware class that
handles it. The midleware is added to the request handler chain.
The request handler returns the success or error xml response directly.
Test refactors to address the changes. Added TestCase class and composer
dependencies.
Php upgraded to 5.6 in order to make brain/monkey library work.
