+++ composer.json
@@ -1,36 +1,45 @@
 {
-  "name": "inpsyde/agb-connector",
-  "description": "Terms & Conditions Connector of IT-Recht Kanzlei",
-  "type": "wordpress-plugin",
-  "license": "GPL-2.0-or-later",
-  "authors": [
-    {
-      "name": "Inpsyde GmbH",
-      "email": "hello@inpsyde.com",
-      "homepage": "https://inpsyde.com",
-      "role": "Company"
+    "name": "inpsyde/agb-connector",
+    "description": "Terms & Conditions Connector of IT-Recht Kanzlei",
+    "type": "wordpress-plugin",
+    "license": "GPL-2.0-or-later",
+    "authors": [
+        {
+            "name": "Inpsyde GmbH",
+            "email": "hello@inpsyde.com",
+            "homepage": "https://inpsyde.com",
+            "role": "Company"
+        }
+    ],
+    "support": {
+        "issues": "https://github.com/inpsyde/agb-connector/issues",
+        "forum": "https://wordpress.org/support/plugin/agb-connector",
+        "wiki": "https://github.com/inpsyde/agb-connector/wiki",
+        "source": "https://github.com/inpsyde/agb-connector"
+    },
+    "require": {
+        "php": ">=5.6",
+        "ext-simplexml": "*",
+        "ext-libxml": "*",
+        "ext-dom": "*",
+        "ext-json": "*"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "~5.7",
+        "brain/monkey": "^2.3",
+        "ptrofimov/xpmock": "^1",
+        "inpsyde/php-coding-standards": "^0.13"
+    },
+    "autoload": {
+        "psr-4": {
+            "Inpsyde\\AGBConnector\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Inpsyde\\AGBConnectorTests\\": "tests/php",
+            "Inpsyde\\AGBConnectorTests\\Unit\\": "tests/php/Unit",
+            "Inpsyde\\AGBConnectorTests\\Functional\\": "tests/php/Functional"
+        }
     }
-  ],
-  "support": {
-    "issues": "https://github.com/inpsyde/agb-connector/issues",
-    "forum": "https://wordpress.org/support/plugin/agb-connector",
-    "wiki": "https://github.com/inpsyde/agb-connector/wiki",
-    "source": "https://github.com/inpsyde/agb-connector"
-  },
-  "require": {
-    "php": ">=5.4",
-    "ext-simplexml": "*",
-    "ext-libxml": "*",
-    "ext-dom": "*",
-    "ext-json": "*"
-  },
-  "require-dev": {
-    "phpunit/phpunit": "~5.7",
-    "inpsyde/php-coding-standards": "^0.13"
-  },
-  "autoload": {
-    "psr-4": {
-      "Inpsyde\\AGBConnector\\": "src/"
-    }
-  }
 }
