+++ src/Middleware/MiddlewareRequestHandler.php
@@ -2,8 +2,9 @@
 
 namespace Inpsyde\AGBConnector\Middleware;
 
-use Exception;
-use Inpsyde\AGBConnector\CustomExceptions\NotSimpleXmlInstanceException;
+use Inpsyde\AGBConnector\CustomExceptions\XmlApiException;
+use Inpsyde\AGBConnector\Plugin;
+use SimpleXMLElement;
 
 /**
  * Class MiddlewareRequestHandler
@@ -47,7 +48,7 @@ class MiddlewareRequestHandler
             ->linkWith(new CheckCredentialsXml())
             ->linkWith(new CheckAuthXml($this->userAuthToken))
             ->linkWith(new CheckTextTypeXml())
-            ->linkWith(new CheckCountryXml())
+            ->linkWith(new CheckCountrySetXml())
             ->linkWith(new CheckTitleXml())
             ->linkWith(new CheckTextXml())
             ->linkWith(new CheckHtmlXml())
@@ -55,7 +56,8 @@ class MiddlewareRequestHandler
             ->linkWith(new CheckPdfFilenameXml())
             ->linkWith(new CheckLanguageXml())
             ->linkWith(new CheckActionXml())
-            ->linkWith(new CheckConfiguration($this->userAuthToken, $this->allocations));
+            ->linkWith(new CheckConfiguration($this->userAuthToken, $this->allocations))
+            ->linkWith(new CheckPostXml($this->allocations));
         return $middleware;
     }
 
@@ -68,14 +70,73 @@ class MiddlewareRequestHandler
     }
 
     /**
-     * @param $xml
+     * @param $data
      *
-     * @return bool|Exception|NotSimpleXmlInstanceException|int
+     * @return string with xml response
      */
-    public function handle($xml)
+    public function handle($data)
     {
-        $response = $this->middleware->process($xml);
+        try {
+            $targetUrl = $this->middleware->process($data);
 
-        return $response;
+            return $this->returnXmlWithSuccess(0, $targetUrl);
+        }catch (XmlApiException $exception){
+
+            return $this->returnXmlWithError($exception);
+        }
+    }
+
+    /**
+     * Returns the XML positive answer
+     *
+     * @param int $code Error code 0 on success.
+     * @param string $targetUrl The url of the site where to find the legal text
+     *
+     * @return string with xml response
+     */
+    public function returnXmlWithSuccess($code, $targetUrl = null)
+    {
+        global $wp_version;
+
+        $xml = new SimpleXMLElement('<?xml version="1.0" encoding="utf-8" ?><response></response>');
+        $xml->addChild('status', 'success');
+        if (!$code && $targetUrl) {
+            $targetUrlChild = $xml->addChild('target_url');
+            $node = dom_import_simplexml($targetUrlChild);
+            $no = $node->ownerDocument;
+            $node->appendChild($no->createCDATASection($targetUrl));
+        }
+        $xml->addChild('meta_shopversion', $wp_version);
+        $xml->addChild('meta_modulversion', Plugin::VERSION);
+        $xml->addChild('meta_phpversion', PHP_VERSION);
+
+        return $xml->asXML();
+    }
+
+    /**
+     * Returns the XML answer with the error
+     *
+     * @param XmlApiException $exception Error code 0 on success.
+     *
+     * @return string with xml response
+     */
+    public function returnXmlWithError($exception)
+    {
+        global $wp_version;
+
+        $xml = new SimpleXMLElement('<?xml version="1.0" encoding="utf-8" ?><response></response>');
+        $xml->addChild('status', 'error');
+        if ($exception) {
+            $xml->addChild('error', $exception->getCode());
+            $messageChild = $xml->addChild('error_message');
+            $node = dom_import_simplexml($messageChild);
+            $no = $node->ownerDocument;
+            $node->appendChild($no->createCDATASection($exception->getMessage()));
+        }
+        $xml->addChild('meta_shopversion', $wp_version);
+        $xml->addChild('meta_modulversion', Plugin::VERSION);
+        $xml->addChild('meta_phpversion', PHP_VERSION);
+
+        return $xml->asXML();
     }
 }
